refactor: split librarySyncEngine into orchestration and diff modules
Extract change detection and incremental fetch logic (detectChanges,
applyChanges, syncPlaylists, syncAlbums) into libraryDiffEngine.ts as
pure functions. librarySyncEngine.ts retains polling loop, tab visibility
handling, listener management, optimistic mutations, and state transitions.
